requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-google-genai
google-generativeai
sentence-transformers
faiss-cpu
huggingface-hub
transformers
torch
pypdf
python-dotenv
chromadb
tiktoken